# backend/src/helpdesk/dto/send-message.dto.ts
import { IsString, IsNotEmpty, MaxLength } from 'class-validator';

export class SendMessageDto {
    @IsString()
    @IsNotEmpty()
    @MaxLength(2000, { message: 'Message is too long' })
    content: string;
}